Align docs plugin with SDK-CLI-Plugin pattern
Mirrors alignment work done in PRs #93-#96 for repo and work plugins.

Changes:
- CLI: Flatten command names with entity prefixes (create → doc-create, types → type-list)
- Plugin: Add 8 CLI-delegating commands for CRUD ops, keep 4 agent-delegating commands
- Agents: Rename to noun form (docs-audit → docs-auditor, docs-write → docs-writer, etc.)
- Config: Update plugin.json to v3.2.0 with new agent paths
- Skills: Update doc-type-selector with new CLI command names

Scope: Functional code only (CLI, commands, agents, skills). Documentation updates deferred.
