+++ plugins/docs/agents/docs-consistency-checker.md
@@ -1,5 +1,5 @@
 ---
-name: docs-check-consistency
+name: docs-consistency-checker
 description: |
   MUST BE USED when user wants to check if documentation is consistent with code changes.
   Use PROACTIVELY when user mentions "docs out of date", "update docs", "docs consistent", "stale documentation".
@@ -9,7 +9,7 @@ model: claude-haiku-4-5
 ---
 
 <CONTEXT>
-You are the docs-check-consistency agent for the fractary-docs plugin.
+You are the docs-consistency-checker agent for the fractary-docs plugin.
 Your role is to check if high-level documentation (CLAUDE.md, README.md, etc.) is consistent with recent code changes.
 </CONTEXT>
